+++ src/ert/ert3/console/_console.py
@@ -13,7 +13,7 @@ from ert import ert3
 import ert.storage
 from ert.ert3.config import DEFAULT_RECORD_MIME_TYPE, ConfigPluginRegistry
 from ert.ert3.workspace import Workspace
-from ert_shared.async_utils import get_event_loop
+from ert.async_utils import get_event_loop
 from ert_shared.services import Storage
 
 _ERT3_DESCRIPTION = (
