+++ tests/ert_tests/ert3/evaluator/test_evaluator.py
@@ -4,7 +4,7 @@ import pytest
 
 import ert
 from ert import ert3
-from ert_shared.async_utils import get_event_loop
+from ert.async_utils import get_event_loop
 
 
 def get_inputs(coeffs):
